v3 2

+++ website/scripts/clean_dir.sh
@@ -1,2 +1,7 @@
 #!/bin/bash
-rm /var/www/html/index.php
+
+isExistFile = 'find /var/www/html/index.php'
+
+if [[ -n $isExistFile ]]; then
+    rm /var/www/html/index.php
+fi
